i3/config: make Teams notifications float and not focus
MS Teams for Linux is not written in a Linux native manner
and therefore does not work well with any window manager, especially
not with tiling ones.

Since MS app is nicer for calls, this change makes the notification
at least not show as a full window. For all other purposes, the
non-official app written by a sole developer is still clearly better
than MSs own.